+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - Requires PHP `8.4`
 - `Innmind\Math\DefinitionSet\Set` is now a final class, all previous implementations are now flagged as internal
+- `Innmind\Math\Monoid\*` are now enums
 
 ### Fixed